add schemas and validateBody
