+++ schema/contacts-schemas.js
@@ -1,0 +1,16 @@
+import Joi from "joi";
+
+
+const constactsAddSchema = Joi.object({
+  name: Joi.string().required().messages({
+    "any.required": "missing required name field",
+  }),
+  email: Joi.string().required().messages({
+    "any.required": "missing required email field",
+  }),
+  phone: Joi.string().required().messages({
+    "any.required": "missing required phone field",
+  }),
+});
+
+export default constactsAddSchema;
